test(shadow-dom-wrapper): enhance tests
- move test file to tests folder
